Bind to all IP address (IPv4 and IPv6) by default.
If the address is 0.0.0.0, the server accepts TCP/IP connections on all server
host IPv4 interfaces.

If the address is ::, the server accepts TCP/IP connections on all server host
IPv4 and IPv6 interfaces. Use this address to permit both IPv4 and IPv6
connections on all server interfaces.

Fixes issue #202.